update remote data source

+++ app/src/main/java/ir/divar/androidtask/data/network/models/request/PostListRequest.kt
@@ -1,3 +1,3 @@
 package ir.divar.androidtask.data.network.models.request
 
-data class PostListRequest(val page: Int, val last_post_date: Int)
+data class PostListRequest(val page: Int, val last_post_date: Long)
